\name{descriptionInfo}
\alias{descriptionInfo}
\alias{biocPkgNameIndex}
\alias{biocViewsIndex}
\alias{manufacturerIndex}
\alias{organismIndex}
\docType{data}
\title{Detailed DESCRIPTION Information}
\description{
  These are the information which will be used to create more detailed
  DESCRIPTION file.
}
\usage{data(descriptionInfo)}
\format{
  It will provide a data frame called "descriptionInfo"  with 7 columns:
  biocPkgName
  organism
  species
  manufacturer
  chipName
  manufacturerUrl
  biocViews
}
\examples{
data(descriptionInfo)
colnames(descriptionInfo)
}
\keyword{datasets}
