\name{pubDataURLs}
\alias{pubDataURLs}
\docType{data}
\title{ Public Domain Data Source URLs}
\description{
  These are the URLs to use in creating a mirror of the public data
  needed to create annotation data packages.
}
\usage{data(pubDataURLs)}
\format{
  The format is:
List of 8
 $ HG  : chr "ftp://ftp.ncbi.nih.gov/pub/HomoloGene/old/"
 $ LL  : chr "ftp://ftp.ncbi.nih.gov/refseq/LocusLink/"
 $ UG  : chr "ftp://ftp.ncbi.nih.gov/repository/UniGene/"
 $ EG  : chr "ftp://ftp.ncbi.nlm.nih.gov/gene/DATA/"
 $ GP  : chr "ftp://hgdownload.cse.ucsc.edu/goldenPath/currentGenomes/"
 $ GO  : chr "http://www.godatabase.org/dev/database/archive/latest/"
 $ YG  : chr "ftp://genome-ftp.stanford.edu/pub/yeast/data_download/"
 $ KEGG: chr "ftp://ftp.genome.ad.jp/pub/kegg/tarfiles/pathway.tar.gz"
}
\examples{
data(pubDataURLs)
}
\keyword{datasets}
